requirements.txt
appium-python-client>=5.2.0
pytest>=9.0.0
pytest-html>=4.1.0
selenium>=4.39.0
